pyproject.toml
[project]
name = "gifttocsv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flashcard-generator>=0.1.1",
    "pygiftparserrgmf>=0.0.5",
]
